requirements.txt
torch
numpy
pandas
scikit-learn
matplotlib
seaborn
librosa
tqdm
umap-learn
jupyter
